Add dependabot and apply pre-commit
