+++ .github/dependabot.yaml
@@ -1,0 +1,15 @@
+# -*- mode: yaml; coding: utf-8 -*-
+#
+# Copyright (C) 2024 Benjamin Thomas Schwertfeger
+# All rights reserved.
+# https://github.com/btschwertfeger
+
+version: 2
+updates:
+  - package-ecosystem: "github-actions"
+    directory: "/"
+    schedule:
+      interval: "monthly"
+    groups:
+      all:
+        dependency-type: "production"
