+++ README.md
@@ -1,17 +1,22 @@
 # [Energy Balance Models - Website](https://btschwertfeger.github.io/Energy-Balance-Models-Website/)
 
-This website includes a simple temperature and a more complex aqua-planet energy balance model. These can be run, then the results are visualized.
+This website includes a simple temperature and a more complex aqua-planet energy
+balance model. These can be run, then the results are visualized.
 
-The goal is to visually represent complex interrelationships by means of simple inputs without programming experience.
+The goal is to visually represent complex interrelationships by means of simple
+inputs without programming experience.
 
-![temperature energy balance model](images/tebm.png)
-![complex energy balance model](images/iebm.png)
+![temperature energy balance model](images/tebm.png) ![complex energy balance
+model](images/iebm.png)
 
 ## Availability
 
 - [https://btschwertfeger.github.io/Energy-Balance-Models-Website/](https://btschwertfeger.github.io/Energy-Balance-Models-Website/)
-- [https://www.awi.de/ ....](https://www.awi.de/fileadmin/user_upload/AWI/Forschung/Klimawissenschaft/Dynamik_des_Palaeoklimas/EnergyBalanceModels/index.html)
+- [https://www.awi.de/
+  ....](https://www.awi.de/fileadmin/user_upload/AWI/Forschung/Klimawissenschaft/Dynamik_des_Palaeoklimas/EnergyBalanceModels/index.html)
 
 ## References
 
-- Wagner, Till & Eisenman, Ian. (2015). How Climate Model Complexity Influences Sea Ice Stability. Journal of Climate. 28. 3998–4014. 10.1175/JCLI-D-14-00654.1.
+- Wagner, Till & Eisenman, Ian. (2015). How Climate Model Complexity Influences
+  Sea Ice Stability. Journal of Climate. 28. 3998–4014.
+  10.1175/JCLI-D-14-00654.1.
